+++ src/main/java/cart/controller/WebPageController.java
@@ -8,6 +8,7 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 
 import cart.controller.dto.ProductResponse;
+import cart.controller.dto.UserResponse;
 import cart.dao.UserDao;
 import cart.dao.dto.UserDto;
 import cart.domain.Product;
