+++ src/main/java/cart/controller/dto/UserResponse.java
@@ -1,4 +1,4 @@
-package cart.controller;
+package cart.controller.dto;
 
 public class UserResponse {
